requirements.txt
streamlit
PyPDF2
langchain-community
openai
python-dotenv
nltk

